Using the update method from bukkit
