+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.coffee</groupId>
     <artifactId>TheWarpPlugin</artifactId>
-    <version>1.2.0</version>
+    <version>1.2.1</version>
     <packaging>jar</packaging>
 
     <name>warpCommand</name>
@@ -72,7 +72,7 @@
         <dependency>
             <groupId>commons-io</groupId>
             <artifactId>commons-io</artifactId>
-            <version>2.11.0</version>
+            <version>2.14.0</version>
         </dependency>
 
         <dependency>
