+++ src/main/resources/plugin.yml
@@ -1,5 +1,5 @@
 name: TheWarpPlugin
-version: '1.2.0'
+version: '${project.version}'
 main: com.coffee.WarpPlugin.Main
 api-version: '1.21'
 authors: [Coffee1307]
